cell03 ex00

+++ cell02/ex00/iszero.py
@@ -1,4 +1,4 @@
-num = int(input())
+num = float(input())
 if num == 0:
     print(f"This number is equal to zero.")
 else:
